requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
diff-match-patch==20230430
Django==5.1.2
django-allauth==65.0.2
django-crispy-forms==2.3
django-import-export==4.1.1
djangorestframework==3.15.2
ezdxf==1.3.3
fonttools==4.54.1
idna==3.10
numpy==2.1.2
pandas==2.2.3
psycopg2==2.9.9
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==11.5.0
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
